+++ packages/awino-ui/src/components/pages/landing/AssetSection/AssetSection.test.tsx
@@ -1,17 +1,12 @@
 import React from 'react';
 
-// import { aboutStats as aboutStatsRecord } from '@/fixtures/about';
 import nextUseRouterMock from '@/mocks/nextUseRouterMock';
 import { cleanup, render } from '@/testing/utils';
 
 import AssetSection from './AssetSection';
 
-const stats = [
-  { value: 89.7, subvalue: 24.72 },
-  { value: 89.7, subvalue: 24.72 },
-  { value: 0.27 },
-  { value: 273.4, subvalue: 52 },
-];
+// export * from '@testing-library/react';
+// export { withThemeRender as render };
 
 beforeAll(() => {
   nextUseRouterMock({
@@ -26,7 +21,10 @@ afterEach(cleanup);
 
 describe('<AssetSection />', () => {
   it('has valid snapshot', () => {
-    const { asFragment } = render(<AssetSection items={stats} />);
+    /*     const { asFragment } = withStore(<AssetSection />, {}, (store: Store) => {
+      store.dispatch(setPageI18nNamespace('market'));
+    }); */
+    const { asFragment } = render(<AssetSection />, {}, 'landing');
     expect(asFragment()).toMatchSnapshot();
   });
 });