+++ packages/awino-ui/src/types/pages/landing.d.ts
@@ -1,0 +1,5 @@
+interface StatsDataItem {
+  value: number;
+  subvalue?: number;
+}
+export type StatsData = StatsDataItem[];